Referenced Typography.Contours and Demo Shared Project to implement glyph rendering Implementing glyph rendering

+++ CSharpMath.SkiaSharp/Font/SkiaGlyphFinder.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System;
 using Typography.OpenFont;
-using TGlyph = System.Int32;
+using TGlyph = System.UInt16;
 
 namespace CSharpMath.SkiaSharp {
   public class SkiaGlyphFinder : IGlyphFinder<TGlyph> {
